gan hoan chinh

+++ Admin/Function/Products/Handle.php
@@ -5,6 +5,8 @@ if (isset($_POST['updatedata'])) {
   $id = $_POST['MaGiay'];
   $TenGiay = $_POST['TenGiay'];
   $GiaBan = $_POST['GiaBan'];
+  $Size = $_POST['Size'];
+  $Màu = $_POST['Màu'];
   $MoTa = $_POST['MoTa'];
   // $NgayCapNhat = $_POST['NgayCapNhat'];
   $SoLuongTon = $_POST['SoLuongTon'];
@@ -37,10 +39,12 @@ if (isset($_POST['updatedata'])) {
     } else {
       if (move_uploaded_file($filename, $path)) {
         if (is_numeric($SoLuongTon) and is_numeric($GiaBan) and is_numeric($GiaBanCu) ) {
-          $query = "UPDATE Giay SET TenGiay='$TenGiay',GiaBan =$GiaBan,MoTa ='$MoTa',AnhBia ='$newname',NgayCapNhat = now(),SoLuongTon =$SoLuongTon,MaLG =$MaLG,MaTH =$MaTH ,MaNCC =$MaNCC ,GiaBanCu =$GiaBanCu   WHERE MaGiay='$id'  ";
+          $query = "UPDATE Giay SET TenGiay='$TenGiay',GiaBan =$GiaBan,MoTa ='$MoTa',AnhBia ='$newname',NgayCapNhat = now(),SoLuongTon =$SoLuongTon,MaLG =$MaLG,MaTH =$MaTH ,MaNCC =$MaNCC ,GiaBanCu =$GiaBanCu,Size = '$Size',Màu = '$Màu'   WHERE MaGiay='$id'  ";
           mysqli_query($conn, $query);
           Edit();
           unlink($path1);
+         
+          
         } else {
           if (!is_numeric($GiaBan)) {
             $loi = "Vui lòng nhập dạng số !";
@@ -62,7 +66,7 @@ if (isset($_POST['updatedata'])) {
     }
   } else {
     if (is_numeric($SoLuongTon) and is_numeric($GiaBan) and is_numeric($GiaBanCu)) {
-    $query = "UPDATE Giay SET TenGiay='$TenGiay',GiaBan =$GiaBan,MoTa ='$MoTa',NgayCapNhat = now(),SoLuongTon =$SoLuongTon,MaLG =$MaLG,MaTH =$MaTH ,MaNCC =$MaNCC  ,GiaBanCu =$GiaBanCu   WHERE MaGiay='$id'  ";
+    $query = "UPDATE Giay SET TenGiay='$TenGiay',GiaBan =$GiaBan,MoTa ='$MoTa',NgayCapNhat = now(),SoLuongTon =$SoLuongTon,MaLG =$MaLG,MaTH =$MaTH ,MaNCC =$MaNCC  ,GiaBanCu =$GiaBanCu,Size = '$Size',Màu = '$Màu'    WHERE MaGiay='$id'  ";
     mysqli_query($conn, $query);
     Edit();}
     else {
@@ -93,8 +97,11 @@ if (isset($_POST['deletedata'])) {
   $AnhBiaCu = $_POST['D_AnhBiaCu'];
   $path1 = "../Images/ImgProducts/". $AnhBiaCu;
   $query = "DELETE FROM Giay WHERE MaGiay='$id'";
+  $query_run_1 = mysqli_query($conn,"DELETE FROM chitietdathang WHERE MaGiay='$id'");
   $query_run = mysqli_query($conn, $query);
-  if ($query_run) {
+  
+
+  if ($query_run and  $query_run_1) {
     unlink($path1);
     Delete();
   } else {
@@ -110,6 +117,8 @@ if (isset($_POST['deletedata'])) {
 if (isset($_POST['insertdata'])) {
   $I_TenGiay = $_POST['I_TenGiay'];
   $I_GiaBan = $_POST['I_GiaBan'];
+  $I_Size = $_POST['I_Size'];
+  $I_Màu = $_POST['I_Màu'];
 
   $I_MoTa = $_POST['I_MoTa'];
   // $I_NgayCapNhat = $_POST['I_NgayCapNhat'];
@@ -150,7 +159,9 @@ if (isset($_POST['insertdata'])) {
           '$I_MaTH',
           '$I_MaNCC',
           1,
-          null
+          null,
+          '$I_Size',
+          '$I_Màu'
           );";
           mysqli_query($conn, $query1);
           Add();
